Make utils module private
