+++ PyPDF2/pdf.py
@@ -67,9 +67,8 @@ from PyPDF2.errors import (  # noqa: F401
     PdfReadWarning,
 )
 
-from . import utils  # noqa: F401
-from .generic import *  # noqa: F401
-from .utils import (  # noqa: F401
+from . import _utils  # noqa: F401
+from ._utils import (  # noqa: F401
     ConvertFunctionsToVirtualList,
     b_,
     formatWarning,
@@ -80,6 +79,7 @@ from .utils import (  # noqa: F401
     str_,
     u_,
 )
+from .generic import *  # noqa: F401
 
 warnings.warn(
     "The PyPDF2.pdf module is deprecated. "
