+++ Tests/test_utils.py
@@ -4,7 +4,7 @@ import os
 
 import pytest
 
-import PyPDF2.utils
+import PyPDF2._utils
 from PyPDF2 import PdfReader
 from PyPDF2.errors import PdfStreamError
 
@@ -18,11 +18,11 @@ RESOURCE_ROOT = os.path.join(PROJECT_ROOT, "Resources")
     [(0, True), (-1, True), (1, True), ("1", False), (1.5, False)],
 )
 def test_isInt(value, expected):
-    assert PyPDF2.utils.isInt(value) == expected
+    assert PyPDF2._utils.isInt(value) == expected
 
 
 def test_isBytes():
-    assert PyPDF2.utils.isBytes(b"")
+    assert PyPDF2._utils.isBytes(b"")
 
 
 @pytest.mark.parametrize(
@@ -37,11 +37,11 @@ def test_isBytes():
     ],
 )
 def test_skipOverWhitespace(stream, expected):
-    assert PyPDF2.utils.skipOverWhitespace(stream) == expected
+    assert PyPDF2._utils.skipOverWhitespace(stream) == expected
 
 
 def test_readUntilWhitespace():
-    assert PyPDF2.utils.readUntilWhitespace(io.BytesIO(b"foo"), maxchars=1) == b"f"
+    assert PyPDF2._utils.readUntilWhitespace(io.BytesIO(b"foo"), maxchars=1) == b"f"
 
 
 @pytest.mark.parametrize(
@@ -54,7 +54,7 @@ def test_readUntilWhitespace():
     ],
 )
 def test_skipOverComment(stream, remainder):
-    PyPDF2.utils.skipOverComment(stream)
+    PyPDF2._utils.skipOverComment(stream)
     assert stream.read() == remainder
 
 
@@ -63,7 +63,7 @@ def test_readUntilRegex_premature_ending_raise():
 
     stream = io.BytesIO(b"")
     with pytest.raises(PdfStreamError) as exc:
-        PyPDF2.utils.readUntilRegex(stream, re.compile(b"."))
+        PyPDF2._utils.readUntilRegex(stream, re.compile(b"."))
     assert exc.value.args[0] == "Stream has ended unexpectedly"
 
 
@@ -71,7 +71,9 @@ def test_readUntilRegex_premature_ending_name():
     import re
 
     stream = io.BytesIO(b"")
-    assert PyPDF2.utils.readUntilRegex(stream, re.compile(b"."), ignore_eof=True) == b""
+    assert (
+        PyPDF2._utils.readUntilRegex(stream, re.compile(b"."), ignore_eof=True) == b""
+    )
 
 
 @pytest.mark.parametrize(
@@ -83,12 +85,12 @@ def test_readUntilRegex_premature_ending_name():
     ],
 )
 def test_matrixMultiply(a, b, expected):
-    assert PyPDF2.utils.matrix_multiply(a, b) == expected
+    assert PyPDF2._utils.matrix_multiply(a, b) == expected
 
 
 def test_markLocation():
     stream = io.BytesIO(b"abde" * 6000)
-    PyPDF2.utils.markLocation(stream)
+    PyPDF2._utils.markLocation(stream)
     os.remove("PyPDF2_pdfLocation.txt")  # cleanup
 
 
@@ -101,12 +103,12 @@ def test_ConvertFunctionsToVirtualList():
 
 
 def test_hexStr():
-    assert PyPDF2.utils.hexStr(10) == "0xa"
+    assert PyPDF2._utils.hexStr(10) == "0xa"
 
 
 @pytest.mark.no_py27()
 def test_b():
-    assert PyPDF2.utils.b_("foo") == b"foo"
-    assert PyPDF2.utils.b_("😀") == "😀".encode("utf-8")
-    assert PyPDF2.utils.b_("‰") == "‰".encode("utf-8")
-    assert PyPDF2.utils.b_("▷") == "▷".encode("utf-8")
+    assert PyPDF2._utils.b_("foo") == b"foo"
+    assert PyPDF2._utils.b_("😀") == "😀".encode("utf-8")
+    assert PyPDF2._utils.b_("‰") == "‰".encode("utf-8")
+    assert PyPDF2._utils.b_("▷") == "▷".encode("utf-8")
